S05.01 - fromSol: removed loadWeatherData
fromSol: suggestions taken over from solution

loadWeatherData was only called from two places, but when the menu
button is used, it is enough to reload the Loader - that's one line, so
we can call that line and move the content of loadWeatherData back to
onCreate

Functionality in onCreate can also be limited. For example, no need to
set the default visibility of views. Apparently initLoader starts the
loading itself, so no need to call startLoading

Also forceLoad was moved to onStartLoading.

Also created a tag() function to append thread ID after TAG. It is not
clear which methods are executing in which order on each thread.